requirements.txt
streamlit
psycopg2-binary
python-dotenv
requests
bs4
